move password strength to password field component
